Avoid row group decoding when possible

+++ core/rust/qdbr/src/cairo.rs
@@ -46,13 +46,13 @@ impl DefaultReturn for u32 {
     }
 }
 
-impl DefaultReturn for i32 {
+impl DefaultReturn for i64 {
     fn default_return() -> Self {
         0
     }
 }
 
-impl DefaultReturn for i64 {
+impl DefaultReturn for u64 {
     fn default_return() -> Self {
         0
     }
